Split up account tests into module
